Dirty always keeps original value of attribute.
This means that changing name from "Foo" to "Bar"
and then back again to "Foo" does no longer count
as a change. Change from "Foo" to "Bar" to "FooBar"
will report "Foo" as name_was too.

This is the behavior which ActiveRecord gives us, but
as far as I managed to see it wasn't exactly what
ActiveModel::Dirty gave us.