Made the ORM's finder forward it's call to the table.
Next up might be to get logic and behavior around a class to
get a connection and to return the correct table object. Then we are
kinda close to be able to do really basic fetching and building object by
the orm.

Still a lot to do though, define DSL for defining table schema,
associations, auto-casting date/time values and other
nice-to-have-features.