Removed the Basic class which inherited from ORM::Base.
This commit will result in a conflict when Vincent merges in his changes
of the finders spec :-( ..but I guess it should be easy to figure out.

Swaped it out for a TestClass. The reason for this is that
I guess we'll never inherit directly from Base, but rather from a Table.
I will also only read attributes from the schema, instead of dynamicly
allow attributes which enters new(attr_hash).