Bugfix: Ensure that all dirty-states are reset when they should.
I noticed a problem when I called reload() and then tried to
set an attribute to something. The dirty module didn't detect
the change, as it thought the value was set back to it's original
value.

I think think this might have been the problem we
encountered with Machinist's make! as well.