Added increment, decrement and atomic_increment to the ORM.
Please note that the atomic_increment is dependent on the adapter, or
wrapper as we currently call it. The Thrift adapter does not currently
support this, or that is; it will do the increment, but not in an atomic
way. When we solve that, or use another wrapper we should be okay
without doing anything with the ORM.