You can now ask fields set, or a column family if an attribute name has been taken.
Now, if the fieldset is contained inside a column family, it will ask
the family for it's opinion if an attribute name has been taken. And,
the plan is next to do so that if a column family is contained inside a
set of families, it will in turn ask the set of families if it knows if
a name has been taken.

This way, any field can ask if it's fields set knows if a name has been
taken, and do so without have to worry if we are part of just that set
of fields, or a set of column families.