Added a add_field_to_column_family method on the table interface.
This method should be used if we dynamically want
to add fields to a class. It's responsible is to
do the things needed for the class to be in a
consistence state. Right now, the only thing it does
is to un-define generated read/write method for
attributes so that next time an_object.an_attribute
is called it will regenerate these methods again to
be sure to include the added method.

Maybe a better solution is simply to ask it to be
generated? I haven't looked at the possibility for 
this yet.