The work on the ColumnInterface has started :-)
I think all previous functionality now is reimplemented in the new
Schema module, and we are ready to implement a ColumnInterface which
will in turn make it possible for us to:

class Address < MassiveRecord::ORM::Column
  field address
  field number, :integer
end

The plan is to simply use Schema::Fields as back-end for keeping the
fields for a column, and to implement DSL methods in the
ColumnInterface as I did with the TableInterface. These interfaces acts
as glue between the ORM class and the schema definitions.