references_many.limit(1) calls find with single id instead of wrapped inside array.
The reason for this is as follows: find(1)
is more explicit than find([1]). If the adapter
is asked to find one record I believe Birger said
that it can do that more efficient with a get-call (?)
instead of opening a scanner, scan one row and close
the scanner. Not 100% sure if it will ever have any
big effect or difference, but it does not hurt to
tell find what we really want to to either, and make
it up to the implementation of find how to optimize
whatever it has to do.