Fixed a problem with find scope raising errors finder options unknown to it.
Reverted and restored functionality of these should be passed on to the
table's finder method. I thought for a second I was kinda smart giving
the user a hint about unknown keys, but it turned out (thank you tests),
that I had forgotten that a lot of different options the scope does not
care about passes through it's way.