"Fixed" bug with is_yaml? returning true if string contains \n--- \n.
...Now, wrapper cell should not do this job at all; it should not
serialize or de-serialize anything, but it should be up to the ORM to do
whatever the column families block dictates.

The Wrapper was written before the ORM and this code comes from the time
when we did not have the ORM. It should really be done something about,
but for now this hotfix will at least minimize the risk for the bug to
arise. Even though; this is a design error as things are now.