Thrift adapter is no longer auto-serializing values.
It will only take care of force string's encoding
written and read via Thrift. The rest will be up to the client
of the adapter, for instance the ORM (which knows
for sure what values to serialize and deserialize).

This change will fix an issue where we guessed if
it was YAML or not based on regular expressions.
The ORM will explicitly do this job now, without
guessing.