String encode nil and null.
String "null" will now be json encoded to "\"null\""
when stored to the database, and nil will be json
encoded to "null", and both will of corse be json
decoded when fetched from the database.

Kinda don't like this commit, as it adds complexity
to strings. Now, if the clients which reads the
database always json decodes every string everything
is fine.

Other solutions, both kinda not optimal:

1: We can serialize _all_ strings,
that will solve our problem, but then we'll end up
with ugly data (double escaped strings) in our database.

2: We can say that string attributes can't be nil,
they will always be saved and restored as empty strings.
We raise an error when strings are nil.