Rewrote how reader methods work.
read_attribute now calls on reader method if it
exists. We prepend the reader method's name with
a _ to mark it as "internal", and it's that
internal method which read_attributes call.

Without this loop it will be impossible for
us to override a reader method, and at the same time
call read_attribute in that method.

Oh, and the reason for making read_attribute
call reader method if it exists is to make it
possible to customize reader methods functionality
at per-attribute level. (See the time zone
conversion module where time attributes are
tailored a bit).