Thrift adapter cell is now storing integers as bytes.
Meaning if you ask it to persist 1 it will store:
\x00\x00\x00\x00\x00\x00\x00\x01
as opposite to before, where it would have stored
the string "1".

This should, if everything goes as planed, allow us
to call atomic increments on these values. We should
also be able to unpack that string back to a Fixnum.