If a Thrift cell has value nil it's mutation will have state isDelete.
This means that we no longer need to go through
the extra hoop of serialize nil and save the
encoded string representation of nil in the ORM.

It also means that we can set an integer to nil,
which was my initial thought when doing this change.
It is a big bonus that we no longer need to
serialize nil.

I don't thing this will have any other implications,
but I guess I'll figure that out soon..