Thrift's cell no longer returns UTF-8 encoded strings.
The reason for this is that we are now making the
adapter work on a bit more "low level", giving it
responsibilities like store Fixnum into binary
strings which Hbase expects.

When the number 222 is read from the database it's
value will actually be '\x00\x00\x00\x00\x00\x00\x00\xDE'
which will raise an 'invalid byte sequence in
UTF-8 (ArgumentError)' when we try to read that
value if we force encode it to UTF-8.

I believe the solution will be that the user of the
adapter (in most cases the ORM) which knows the
schema will know when to force encode UTF-8 (on
strings and serialized hashes etc) and when not to
(on obviously integers).