Schema::Field no longer encodes nil.
It's the adapter's responsibility to remove cells
which has a nil-value.

When the record is read from the database
again it's values will no
longer include any value for that attribute set
to nil, thus the nil-value is persisted/represented by
the fact that the schema knows about the attribute
but the record has no such attribute persisted in
the database.