Fixed a STI related problem loading class. Ex:
You have a class Person and Friend which subclasses Person.
You store one Person in the database. It's type will be nil
(as we don't persist class when storing base_class).

Doing a Friend.first did return a Friend, which in
fact it should have been instantiated as a Person.