records_starts_from proc can override finder method + take a block.
This is ground work for make references_many_proxy'
find_in_batches() work as it will call the proc,
override the finder method to be find_in_batches,
and pass it's proc on to the MassiveRecord::ORM::Base'
find_in_batches method. As usual the proc sets the
:start option which should ensure correct scoping
as before.