Raises error if you try to do first() on subclass (STI). Fixes #50.
This is done due to the fact that we had a bug given you had:
A Person class, and a sub class Friend. If one Person had id 1, and
Friend had id 2 and you did a Friend.first you got back nil.

In order to solve this we can do a Person.all and select every Friend,
then to first on the array. Or we can do a more Person.find_in_batches
and break when we find the first Friend. Either way; it is sadly
inefficient.

The final solution to this is to do the filtering in hbase itself, but
this can't, as far as I know, be done in a dynamic way which we need,
through Thrift. Thus this will be adapter dependent feature when someone
implements JRuby adapter using Java API.