IdentityMap will not add records which are loaded with :select option.
If :select is given it means that they are not fully
loaded (or in fact, it means that they _might not_
be fully loaded), thus we should not be returned
this object if we do a find() the second time with
no :select option.

(At a later point we might have a state telling if
the record is partially loaded or not. For instance
if we :select => [:all, :families, :the, :class, :has]
then we should persist the object in the IdentityMap.)