A references_many answers to all(finder_options).
This fixes #55, so you can do a_person.cars.all(:offset => "id-2").

I'd like to refactor this proxy a bit; as it now
has multiple responsibilities. It handles its own
proxy work and delegation, but it also should be
able to better keep state on the scope the client
asks of it.

For instance, if you do person.cars.limit(2) you'll
get back 2 records right away. It would be nice to
have a smarter scope so we can do
person.cars.limit(2).offset("id-2"). We have
this functionally on class finders so its a bit
shame that its not on relations.