Embeds many proxy has a simple @raw which is a hash keeping raw attributes.
This is a very wip commit!

Plan is to populate this hash with raw data when we
do a find from parent document. Then we load the
proxy's object from the raw data only when we need
to.

A lot of stuff to do here though. To mention a few:
- If @raw is not loaded (due to not selecting the column family)
  we need to do a db query.
- @raw might be renamed @raw_from_database (?).
- We might need to validate records, like references many do.
- We need to set added record's inverse attribute correctly
  (meaning that if a car is embedded in a person the
  car needs to get it's person attribute set).
- Car needs to have validates presence of parent doc.
- When you do car.save it should call save on parent.