Embeds many proxy returns hash which will be mixed in to table update in parent.
It means that when parent is saved it will have to
go through all of it's embedded proxies and ask
if it has any changes which it wants to have done.

If it has, it will read changes from this method
and send it on to to adapter's row mutations.