Added polymorphic embedded in.
It's behavior is exactly equal to the embedded in,
except that it is not checking the target's type.

I could have made it via a simple if-statement, but
I think this is better; keeping the same structure
as we already have and in case of we need to extend
it with custom logic we are free to do so.