Added tests for timestamps on embedded records.
Kind of a problem here, when we add records to an
embeds many collection we are saving it (through
the parent after every record has been added), but
we are hooking into internals of the embedded records
setting for instance @new_record to false etc,
instead of just calling save on them. (Cos we kinda
don't want to save them (again), it was just done).

So either we'll have to keep hacking in the
embeds_many_proxy.parent_has_been_saved! or we'll
need to clean this up now, which is a lot better.