Fixed order of attributes assignment.
We had a situation where

record.attributes = {
  :embedded_in => record_with_embeds_many,
  :name => 'foo'
}

didn't set record inside of record_with_embeds_many,
since it only (the embeds_many collection) only
accepts valid records. And if record had no name
before attributes assignment it would try to assign
embedded_in before setting it's name; in which
case it will not be valid.

Now, you might argue that record should still be
accepted in to the embeds_many collection, but
for now its like this to keep it a bit simpler.