Added database_id to ORM::Embedded.
Will be used when storing embedded records in to
column family; in a way to namespace embedded records.
It gives us the ability to mix embedded collections
and embedded collections with "normal" attributes.

..I think it is cleaner (and saver) to have a
dedicated column family per embedded collection, but
Companybook wanted to have this ability available
to developers, and as far as I understand it Hbase
itself don't like having too many column families
(ref resource on Companybook).