Silently ignore multi param on invalid format, setting value to nil.
I don't know if we need to take care of this in
another way, but as far as I see now; if the user
chooses to manipulate the values sent via Rails'
form date/time helpers, then we just don't care
and we do set the value to nil silently.

This should complete #71.