Callbacks on embedded when owner.save().
The before- and after validation is ran cos
parent will validate associated. Thus, we don't
need to run validations in the embeds_many proxy's
parent_will_be_saved!

Save is stilled applied, so the rest of the
callbacks are applied thereafter.

Now, there is a problem with embedded before-callbacks.
Normally, returning false inside of one will short
circuit the save call, but in this case it does not.